requirements.txt
fastapi
uvicorn
mysql-connector-python
pandas==2.2.3
openpyxl==3.1.5
python-multipart==0.0.10
tensorflow
keras
scikit-learn
matplotlib
PyJWT